inline plot_segments into graph
